requirements.txt
pdfminer.six>=20221105
openpyxl>=3.0.9
unidecode>=1.3.4
pandas>=1.4.2
customtkinter>=5.0.0
pillow>=9.0.0
tkcalendar>=1.6.1
tkinter-tooltip>=2.1.0
requests>=2.27.1
python-dateutil>=2.8.2